Move the search view into the molo.core module
