+++ molo/core/cookiecutter/scaffold/{{cookiecutter.directory}}/{{cookiecutter.app_name}}/urls.py
@@ -9,7 +9,6 @@ from wagtail.wagtailadmin import urls as wagtailadmin_urls
 from wagtail.wagtaildocs import urls as wagtaildocs_urls
 from wagtail.wagtailcore import urls as wagtail_urls
 
-from .views import search
 
 # implement CAS URLs in a production setting
 if settings.ENABLE_SSO:
@@ -27,7 +26,6 @@ urlpatterns += patterns(
     url(r'^django-admin/', include(admin.site.urls)),
     url(r'^admin/', include(wagtailadmin_urls)),
     url(r'^documents/', include(wagtaildocs_urls)),
-    url(r'search/$', search, name='search'),
 
 {% for app_name, regex in cookiecutter.include %}
     url(r'{{regex}}',
