--- molo/core/cookiecutter/scaffold/{{cookiecutter.directory}}/{{cookiecutter.app_name}}/views.py
@@ -1,42 +1,0 @@
-from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.shortcuts import render
-from django.utils.translation import get_language_from_request
-
-from molo.core.utils import get_locale_code
-from molo.core.models import ArticlePage
-from wagtail.wagtailsearch.models import Query
-
-
-def search(request, results_per_page=10):
-    search_query = request.GET.get('q', None)
-    page = request.GET.get('p', 1)
-    locale = get_locale_code(get_language_from_request(request))
-
-    if search_query:
-        results = ArticlePage.objects.filter(
-            languages__language__locale=locale
-        ).values_list('pk', flat=True)
-        # Elasticsearch backend doesn't support filtering
-        # on related fields, at the moment.
-        # So we need to filter ArticlePage entries using DB,
-        # then, we will be able to search
-        results = ArticlePage.objects.filter(pk__in=results)
-        results = results.live().search(search_query)
-
-        Query.get(search_query).add_hit()
-    else:
-        results = ArticlePage.objects.none()
-
-    paginator = Paginator(results, results_per_page)
-    try:
-        search_results = paginator.page(page)
-    except PageNotAnInteger:
-        search_results = paginator.page(1)
-    except EmptyPage:
-        search_results = paginator.page(paginator.num_pages)
-
-    return render(request, 'search/search_results.html', {
-        'search_query': search_query,
-        'search_results': search_results,
-        'results': results,
-    })
