+++ molo/core/urls.py
@@ -1,8 +1,11 @@
 from django.conf.urls import patterns, include, url
 
+from .views import search
+
 
 urlpatterns = patterns(
     '',
+    url(r'search/$', search, name='search'),
     url(
         r'^locale/(?P<locale>[\w\-\_]+)/$',
         'molo.core.views.locale_set',
